Refactor camera activity
